Split more `no0` enemy entities for Ctulhu and Stone Rose
