+++ config/splat.us.stno0.yaml
@@ -78,7 +78,8 @@ segments:
       - [0x41624, .rodata, e_plate_lord]
       - [0x416F0, .rodata, e_marionette]
       - [0x4171C, .rodata, e_slinger]
-      - [0x41738, .rodata, 57C20]
+      - [0x41738, .rodata, e_stone_rose]
+      - [0x41760, .rodata, e_ctulhu]
       - [0x4178C, .rodata, e_axe_knight]
       - [0x417B0, .rodata, e_ouija_table]
       - [0x41804, .rodata, e_flea_man]
@@ -111,7 +112,8 @@ segments:
       - [0x55E4C, c, e_ghost]
       - [0x56254, c, e_marionette]
       - [0x575E4, c, e_slinger]
-      - [0x57C20, c, 57C20]
+      - [0x57C20, c, e_stone_rose]
+      - [0x59264, c, e_ctulhu]
       - [0x5AEBC, c, e_axe_knight]
       - [0x5BDCC, c, e_ouija_table]
       - [0x5C64C, c, e_flea_man]
