+++ config/symbols.us.stno0.txt
@@ -12,6 +12,7 @@ g_EInitElevator = 0x80180B60;
 g_EInitGhostEnemy = 0x80180B9C;
 g_EInitSlinger = 0x80180BB4;
 g_EInitSlingerPieces = 0x80180BC0;
+g_EInitSlingerRib = 0x80180BCC;
 g_EInitAxeKnight = 0x80180BFC;
 D_80180C6A = 0x80180C02;
 g_EInitAxeKnightAxe = 0x80180C08;
@@ -189,6 +190,7 @@ EntityGhostEnemy = 0x801D5E4C;
 EntityGhostEnemySpawner = 0x801D606C;
 EntitySlinger = 0x801D7670;
 EntitySlingerPieces = 0x801D7B40;
+EntitySlingerRib = 0x801D7C20;
 func_801C3F9C = 0x801DAEBC;
 func_801C4198 = 0x801DB0B8;
 func_801C4550 = 0x801DB470;
